+++ tools/dbms/restore.do.pl
@@ -151,10 +151,18 @@ sub doRestore {
 			});
 		}
 		# honor --monitor option
+		# record the topic and the payload to be able to display them in a grafana dashboard
 		if( $opt_monitor ){
-			my $topic = $objNode->name().'/executionReport/'.$ep->runner()->command().'/'.$ep->runner()->verb()."/$opt_service/$opt_database/$mode";
-			my $payload = $mode eq 'full' ? $opt_full : $opt_diff;
-			$res = TTP::commandExec( "dbms.pl sql -service MonitorDB -command \"delete from Monitor.dbo.RESTORES where topic='$topic'\"" );
+			my $root = $objNode->name().'/executionReport/'.$ep->runner()->command().'/'.$ep->runner()->verb()."/$opt_service/$opt_database";
+			my $topic = "$root/$mode";
+			my $payload;
+			if( $mode eq 'full' ){
+				$res = TTP::commandExec( "dbms.pl sql -service MonitorDB -command \"delete from Monitor.dbo.RESTORES where topic like '$root/%'\"" );
+				$payload = $opt_full;
+			} else {
+				$res = TTP::commandExec( "dbms.pl sql -service MonitorDB -command \"delete from Monitor.dbo.RESTORES where topic='$topic'\"" );
+				$payload = $opt_diff;
+			}
 			$res = TTP::commandExec( "dbms.pl sql -service MonitorDB -command \"insert into Monitor.dbo.RESTORES ( topic, payload ) values ( '$topic', '$payload' )\"" );
 		}
 	}