+++ ChangeLog.md
@@ -10,6 +10,7 @@
 	- Fix MongoDB connection checks
 	- dbms.pl restore has '--monitor' option, thus bumping minor candidate version number
 	- MongoDB now accepts (some) NoSql commands
+	- MongoDB is able to translate (some) SQL commands
 
 ### 4.21.1